Refactor and support multiple alarms files
